Renamed to topic transforms and refactored
