--- onboard/catkin_ws/src/system_utils/launch/pose_transforms.launch
@@ -1,3 +1,0 @@
-<launch>
-    <node name="pose_transforms" pkg="system_utils" type="pose_transforms.py" output="screen"></node>
-</launch>
